added routing to the navbar and implemented the doodle board on our Todoodle page.
