+++ package.json
@@ -11,8 +11,10 @@
     "react": "^17.0.2",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.2",
+    "react-drawing-board": "^0.3.6",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-sticky-nav": "^4.0.6",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
