+++ src/App.css
@@ -1,10 +1,20 @@
-img {
-  height: 50vh;
-  width: auto;
+.carousel {
+  align-content: center;
+  margin: 20px;
+  max-width: 100%;
+  max-height: 100%;
 }
 
-.carousel {
-  margin: 50px;
+
+input {
+  text-align: center;
+  height: 30px;
+  width: 150px;
+  align-self: center;
+}
+
+.todoapp {
+  align-items: center;
 }
 
 /*
