+++ src/components/TodoItem.js
@@ -1,5 +1,5 @@
 import { useContext } from "react"
-import { TodosContext } from "../App";
+import { TodosContext } from "../views/Home";
