+++ src/views/Todoodle.js
@@ -1,0 +1,15 @@
+import Youtube from "../components/Youtube"
+import "../App.css";
+import DrawingBoard from 'react-drawing-board';
+  
+  function Todoodle(){
+        
+  return (
+    <>
+        <Youtube/>
+        <DrawingBoard />
+    </>
+  )
+}
+
+export default Todoodle;
